pyproject.toml
[tool.poetry]
name = "lunav2"
version = "0.1.0"
description = ""
authors = ["Kruril <guillaume.charlier.2017@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
google-api-python-client = "^2.22.0"
google-auth-oauthlib = "^0.4.6"
boto3 = "^1.20.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
